Restructure pages folder
