+++ react/pages/address-create/AddressCreate.js
@@ -1,0 +1,68 @@
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
+import { graphql } from 'react-apollo'
+import { compose, branch, renderComponent, withProps } from 'recompose'
+import Header from '../shared/BaseHeader'
+import ErrorAlert from '../../components/shared/ErrorAlert'
+import AddressFormBox from '../../components/Addresses/AddressFormBox'
+import GetName from '../../graphql/getName.gql'
+
+class AddressCreate extends Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      shouldShowError: false,
+    }
+  }
+
+  goBack = () => {
+    this.props.history.push('/addresses')
+  }
+
+  handleError = () => {
+    this.setState({ shouldShowError: true })
+  }
+
+  dismissError = () => {
+    this.setState({ shouldShowError: false })
+  }
+
+  render() {
+    const { profile } = this.props
+    const { shouldShowError } = this.state
+
+    return (
+      <section>
+        <Header
+          titleId={'pages.addressCreate'}
+          backButton={{ id: 'pages.addresses', path: '/addresses' }}
+          shouldAlwaysShowBackButton
+        />
+        <main className="mt6">
+          {shouldShowError && (
+            <div className="mb6 mw6 pr5-ns">
+              <ErrorAlert onDismiss={this.dismissError} />
+            </div>
+          )}
+          <AddressFormBox
+            isNew
+            onAddressSaved={this.goBack}
+            onError={this.handleError}
+            profile={profile}
+          />
+        </main>
+      </section>
+    )
+  }
+}
+
+AddressCreate.propTypes = {
+  profile: PropTypes.object.isRequired,
+}
+
+const enhance = compose(
+  graphql(GetName),
+  branch(({ data }) => data.loading, renderComponent(null)),
+  withProps(({ data }) => ({ profile: data.profile })),
+)
+export default enhance(AddressCreate)
